+++ ddth-tsc-cassandra/src/test/java/qnd/tsc/cassandra/QndCacheTest.java
@@ -1,0 +1,5 @@
+package qnd.tsc.cassandra;
+
+public class QndCacheTest {
+
+}